+++ jabgui/src/main/java/org/jabref/gui/frame/JabRefFrameViewModel.java
@@ -20,11 +20,11 @@ import javafx.beans.value.ObservableBooleanValue;
 import javafx.scene.control.ButtonType;
 
 import org.jabref.cli.CliImportHelper;
-import org.jabref.gui.ClipBoardManager;
 import org.jabref.gui.DialogService;
 import org.jabref.gui.LibraryTab;
 import org.jabref.gui.LibraryTabContainer;
 import org.jabref.gui.StateManager;
+import org.jabref.gui.clipboard.ClipBoardManager;
 import org.jabref.gui.importer.ImportEntriesDialog;
 import org.jabref.gui.importer.actions.OpenDatabaseAction;
 import org.jabref.gui.preferences.GuiPreferences;